Move Show related methods to respective classes
