+++ src/main/java/com/noticemedan/cinema/Application.java
@@ -1,17 +1,15 @@
 package com.noticemedan.cinema;
 
 import com.noticemedan.cinema.entity.ShowEntity;
-import com.noticemedan.cinema.service.CustomerService;
-import com.noticemedan.cinema.service.TestService;
+import com.noticemedan.cinema.service.ShowService;
 
-import java.util.Arrays;
 import java.util.List;
 
 public class Application {
     public static void main(String[] args) {
 
-        TestService testService = new TestService();
-        List<ShowEntity> shows = testService.getShows();
+        ShowService showService = new ShowService();
+        List<ShowEntity> shows = showService.getShows();
 
         System.out.println("Hey");
     }
